Link torch_python directly
libinterpreter needs its own copy of `torch_python`, which links against
its own copy of `libpython`. Before, we achieved this by directly adding
a whitelist of built objects as sources to the `libinterpreter` target.
This is fragile, since we need to edit it every time someone alters
`torch_python`.

This commit instead an object library (`torch_python_obj`). This is used
in two ways:
1. The regular Python extension (`libtorch_python`) links against it and
`libtorch` to produce a shared library.
2. `libinterpreter` statically links against it.

Also needed to fix a static destructor issue in our PassManager.